scrapers/requirements.txt
geojson
pymongo
asyncio
aiohttp
beautifulsoup4
requests
python-dateutil